Move signal generators to a separate file
