+++ src/franc/evaluation/__init__.py
@@ -14,10 +14,12 @@ from .metrics import (
 )
 from .evaluation import (
     EvaluationRun,
-    TestDataGenerator,
     residual_power_ratio,
     residual_amplitude_ratio,
     measure_runtime,
+)
+from .signal_generation import (
+    TestDataGenerator,
 )
 from .report_generation import ReportElement
 from .filter_interface import FilterInterface, make_2d_array, handle_from_dict
